Move book promo into WIP section in Storybook
