+++ cardigan_beta/.storybook/config.js
@@ -7,10 +7,12 @@ function loadStories() {
   const pages = require.context('../stories/pages', true, /\.js$/);
   const components = require.context('../stories/components', true, /\.js$/);
   const global = require.context('../stories/global', true, /\.js$/);
+  const wip = require.context('../stories/wip', true, /\.js$/);
 
   pages.keys().forEach((filename) => pages(filename));
   components.keys().forEach((filename) => components(filename));
   global.keys().forEach((filename) => global(filename));
+  wip.keys().forEach((filename) => wip(filename));
 }
 
 addDecorator(withKnobs);
